[bug] always initialize sync db
Both deferred encryption and decryption rely on a special sync db. Previous to
this fix, the sync db was only initialized if a syncer was configured with
deferred encryption capabilities. This was a problem when the syncer was not
configured like so, but the actual sync method was initiated configured to do
deferred decryption.

This commit fixes this by always initializing the sync db, so we have the
option of doing all combinations of deferred encryption and decryption.