[feature] use twisted.web.client in client sync
This change uses twisted deferreds for the whole syncing process and paves the
way to implementing other transport schemes. It removes a lot of threaded code
that used locks and was very difficult to maintain, and lets twisted to the
dirty work. Furthermore, all blocking network i/o is now handled
asynchronously by the twisted.

This commit removes the possibility of interrupting a sync, and we should
reimplement it using cancellable deferreds if we need it.